requirements.txt
torch==2.7.0+cu128
torchvision==0.22.0+cu128
torchaudio==2.7.0+cu128

# Core dependencies
ultralytics==8.2.75
flask==3.1.2
waitress==3.0.2
opencv-python==4.10.0.84
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.1
tqdm==4.66.5
imageio==2.35.1

# Additional dependencies
pyyaml==6.0.1
requests==2.32.3
matplotlib==3.9.1.post1
pandas==2.2.2
scipy==1.14.0
seaborn==0.13.2

# Utilities
psutil==6.0.0
tabulate==0.9.0
